requirements.txt
# NEO GOD Ultra Framework Requirements v2.0
# Python >=3.10 필수

# ========================================
# 고속 Excel 엔진 (핵심 개선!)
# ========================================
python-calamine>=0.6.0      # Rust 기반 초고속 Excel 리더 (10-18배 가속)

# ========================================
# Google Cloud BigQuery
# ========================================
google-cloud-bigquery>=3.11.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# ========================================
# 데이터 처리
# ========================================
pandas>=2.2.0               # Calamine 엔진 내장 지원 (중요!)
openpyxl>=3.1.0             # 수식 추출용 폴백
numpy>=1.24.0
xlrd>=2.0.1
msoffcrypto-tool>=6.0.0     # Excel 파일 암호화 확인용

# ========================================
# Parquet 지원
# ========================================
pyarrow>=14.0.0             # Parquet + Snappy 압축

# ========================================
# 유틸리티
# ========================================
psutil>=5.9.0               # 메모리 모니터링
pyyaml>=6.0                 # 설정 파일
python-dateutil>=2.8.2      # 날짜 처리
